Checkpoint in implementing CJ minimization, A_out should be done
Also some nasty bugs with openmp fixed, reducing over fortran arrays seemed to add some noise to the results, of the order ~1.d-8
The code looked ok to me, but the problem surfaces with bot gfortran 4.8 and ifort, so maybe it is not a compiler bug. Or maybe openmp implementations in fortran sucks in general, which may be.

q_grid.f90 renamed to q_grids.f90 for fun

some stuff movedaroudn a bit