A few quite important bug fixes thanks to Francesco Macheda who helped with debugging.
A factor 1/2 was missing in the SMA thermal conductivity i.e. the intrinsic HWHM was used instead of the intrinsic HWHM. On the other hand, the correct FWHM was used for the isotopic and casimi contribution, now all three values are consistent and, more importantly correct.

The CGP algorithm, being checkd on SMA, was also missing a factor 0.5