Tentative try to get the same result with nbnd_occ = nbnd and without, also without breaking the code in the insulating case when nbnd>nelec/2
I'm relatively convinced that the correct nbnd_occ to use in solve linter is the one at ik_wfc (i.e. at k) and not at ik_prj (i.e. k-q in the simple case), but I'm not 100% sure. However, it does go much better like this

Also, it is still not perfect, and there are more residualse in the FFT when using this with respect to using nbnd_occ=nbnd