Added two input options optimize_grid optimize_grid_thr
Set the first to true and the integratio grid fro the self-enenrgy will be analyzed and optimized by throwing away the points that do not contribute to the final result. THis is done by pre-computing the linewidth using a constant D3, and saving only the point required to get the total to an accuracy optimize_grid_thr (i.e. optimize_grid_thr=0.01 -> 99%)

Currently only implemented for linewidth along a line, and very inefficient in parallel