Added the mu_initial input parameter for the calculation of final state
When mu_initial is specified, e_initial is automatically set to the energy of the mu-th eigenvalue, unless specified otherwise. In this case, the additional columns of the file will contain the decomposition over the mode of the final state.

If mu_initial is not set, than e_initial must be specified. Note that in this case, the code will compute the final state decomposition of the spectra weight at e_initial, and the additional columns in the file will include the contribution of each modeto the spectra weight. This is quite different from above! 

I.e. if you do not set mu_initial that the column of state NU will be equal to the "total" if computed with mu_initial=NU